Feat: CommonButton 컴포넌트 변경에 따른 사용법 수정
